+++ adapters/fixnum/example/example.dart
@@ -1,6 +1,5 @@
 library json_mapper_fixnum.example;
 
-import 'package:dart_json_mapper/annotations.dart';
 import 'package:dart_json_mapper/json_mapper.dart';
 import 'package:dart_json_mapper_fixnum/dart_json_mapper_fixnum.dart';
 import 'package:fixnum/fixnum.dart' show Int32;