+++ adapters/mobx/lib/dart_json_mapper_mobx.dart
@@ -1,9 +1,6 @@
 library json_mapper_mobx;
 
-import 'package:dart_json_mapper/annotations.dart';
-import 'package:dart_json_mapper/converters.dart';
 import 'package:dart_json_mapper/json_mapper.dart';
-import 'package:dart_json_mapper/type_info.dart';
 import 'package:mobx/mobx.dart';
 
 final mobXTypeInfoDecorator = MobXTypeInfoDecorator();