+++ mapper/example/example.dart
@@ -1,7 +1,3 @@
-library json_mapper.example;
-
-import 'package:dart_json_mapper/annotations.dart';
-import 'package:dart_json_mapper/converters.dart';
 import 'package:dart_json_mapper/json_mapper.dart';
 
 import 'example.reflectable.dart'; // Import generated code.