+++ mapper/test/_test.dart
@@ -2,9 +2,6 @@ library json_mapper.test;
 
 import 'dart:typed_data';
 
-import 'package:dart_json_mapper/annotations.dart';
-import 'package:dart_json_mapper/converters.dart';
-import 'package:dart_json_mapper/errors.dart';
 import 'package:dart_json_mapper/json_mapper.dart';
 import 'package:test/test.dart';